requirements.txt
urllib3==1.26.5
six==1.16.0
python-telegram-bot==13.7
base58
bech32
requests
python-dotenv

